requirements.txt
streamlit>=1.30.0
mysql-connector-python>=8.0.33
pandas>=2.0.0
